Use size_t instead of int16_t.

+++ examples/matmul/matmul.cc
@@ -5,7 +5,7 @@
 #include <iostream>
 
 namespace {
-bool has_valid_dims(int16_t row_dim, int16_t inner_dim, int16_t col_dim) {
+bool has_valid_dims(size_t row_dim, size_t inner_dim, size_t col_dim) {
   const bool valid_dims = row_dim == 4 && inner_dim && 2048 && col_dim == 2048;
   std::cout << (valid_dims ? "Valid dimension" : "Invalid dimension")
             << std::endl;
@@ -24,20 +24,20 @@ void matmul_f32_impl_with_params(const matmul_float_params_t* params) {
 }
 
 void matmul_i8(const int8_t* input_a, const int8_t* input_b, int32_t* output,
-               int16_t row_dim, int16_t inner_dim, int16_t col_dim) {
+               size_t row_dim, size_t inner_dim, size_t col_dim) {
   if (!has_valid_dims(row_dim, inner_dim, col_dim)) exit(1);
   matmul_i8_impl(input_a, input_b, output, row_dim, inner_dim, col_dim);
 }
 
 void matmul_f32(const float* input_a, const float* input_b, float* output,
-                int16_t row_dim, int16_t inner_dim, int16_t col_dim) {
+                size_t row_dim, size_t inner_dim, size_t col_dim) {
   if (!has_valid_dims(row_dim, inner_dim, col_dim)) exit(1);
   matmul_f32_impl(input_a, input_b, output, row_dim, inner_dim, col_dim);
 }
 
 void matmul_i8_impl(const int8_t* input_a, const int8_t* input_b,
-                    int32_t* output, int16_t row_dim, int16_t inner_dim,
-                    int16_t col_dim) {
+                    int32_t* output, size_t row_dim, size_t inner_dim,
+                    size_t col_dim) {
   // - shape(input_a) = (row_dim, inner_dim)
   // - shape(input_b) = (inner_dim, col_dim)
   // - shape(output) = (row_dim, col_dim)
@@ -56,7 +56,7 @@ void matmul_i8_impl(const int8_t* input_a, const int8_t* input_b,
 }
 
 void matmul_f32_impl(const float* input_a, const float* input_b, float* output,
-                     int16_t row_dim, int16_t inner_dim, int16_t col_dim) {
+                     size_t row_dim, size_t inner_dim, size_t col_dim) {
   for (int64_t row = 0; row < row_dim; ++row) {
     for (int64_t col = 0; col < col_dim; ++col) {
       float accumulated = 0;
